refactor(storage): make some code more testable
This just changes `storage::internal::RetryObjectReadSource` to make the
class more testable. The backoff function is provided as an argument, so
it can be mocked. Likewise, the code to start new downloads is provided
by a factory function.

Incidentally, I captured the current options using
`internal::ImmutableOptions`, saving a number of copies in the process.