fix(storage): only backoff before resume attempts
`storage::Client::ReadObject()` resumes a download that gets
interrupted (controlled by policy). On the first resume attempt, the
library does not back off (sleep), becuase there is no reason to believe
the problem is load related. If the first resume fails, the library
backsoff before each attempt, as the problem might be load related after
this point.

The library was *also* backing off before issuing the first `Read()` on
the newly created source of data. That effectively doubles the backoff
time, and leaves the resumed connection idle for (potentially) a long
time when there are multiple resume attempts needed.