impl(ACv2): `OpenObject()` performs the first read (#47)
`OpenObject()` needs to perform the first `Read()` call or we may run
into infinite retry / resume loops.  This function used to create a
streaming RPC, call `Start()`, and then call `Write()` to send the
initial request. On errors it would call `Finish()`. This is not enough
to detect if the service accepted the request. A successful `Write()`
only indicates that the request was **sent**, we need to wait for the
first `Read()` response to determine if the service accepted the
request.