cleanup(storage): simplify integration test init
This is the first step to clean up the integration test initialization.
It removes unused code in the common initialization function, avoids
using `StatusOr<>` when it makes no sense, and prefers using `Options{}`
instead of multiple overloads.

I stopped here because `MakeBucketIntegrationTestClient()` is not used
in as many places and I think it will make this change easier to grok.