fix(GCS+gRPC): handle offsets in `Hash*::Update()`
When reading data from GCS we don't necessarily know the initial offset.
Reading the last N bytes is expressed as a negative offset in the
request, and the returned offset is not known until the first message.
But once the offset is known, receiving out of order data indicates a
bug in the client library, and should be detected.

The code was assuming the initial offset was always 0.