fix(GCS+gRPC): reads resume only after some data (#14579)
By default we want to resume downloads on any error, as long as at least
some data was received. Even if the final status of the stream is scary
(say `NOT_FOUND`), we know the previous attempt was working. Seems
worthwhile to attempt a resume in this case. We also don't want to
believe a resume was "successful" until at least some data was received.
We don't get errors until the first `Read()` attempt.