emulator: tests: exclude "programs" from tests. (#5)
"programs" are not tests in general and can be server binaries which
wait forever and cause the entire test suite to hang and/or
timeout.

In particular add_test should only be called for binaries that are
tests, not arbitrary programs.

TESTED=Verified that "make test" no linger tries to run (and therefore
hang on) the emulator binary.