fix: emulator: fix MSVC compilation error due to storing a int64_t in a size_t.
The compiler spits out an error in pedantic mode because on 32 bit
platforms, there is a usually the possibility of an overflow. However,
the quantity is guaranteed to be 32
bit (https://github.com/googleapis/googleapis/blob/74657e8a6690b249c048f685124ee3b8473b70b4/google/bigtable/v2/data.proto#L497)
so use int32_t to store it and fix the MSVC compiler warning and failure.