fix: make googletest a regular dependency
The googletest dependency was previously marked as a dev_dependency,
which prevented it from being propagated to downstream users. This
caused build failures for users of our mock libraries, which have a
public dependency on googletest.

This change removes dev_dependency to make the mock libraries usable
for consumers of our Bzlmod module.