feat: add local bzlmod consumer test
This commit adds a minimal Bazel project that consumes google-cloud-cpp
as an external dependency using Bzlmod.

The test is designed to reproduce the canonical repository name mismatch
reported in b/430085995 and will be used to verify that the changes
in this pull request resolve the issue from a downstream user's
perspective.