fix(bazel): update grpc checksum
The gRPC archive at v1.69.0 seems to have been updated, as the checksum
of the downloaded file no longer matches the value in the WORKSPACE and
MODULE.bazel files.

This change updates the sha256 and integrity values to match the new
checksum of the gRPC archive. This fixes the checksum mismatch error
for both bzlmod and non-bzlmod builds.