fix(bazel): force re-download of grpc archive
The CI builds are failing with a patch failure when trying to patch the
gRPC archive. This is because the CI machines are using a cached version
of the archive that is either corrupted or outdated.

This change forces Bazel to re-download the gRPC archive by adding a
dummy query parameter to the URL. This will bypass the cache and ensure
that the correct version of the archive is used.