fix(pubsub): fix signed-unsigned comparison in test
The `windows/cmake-release-x86-presubmit` build is failing with a
signed/unsigned comparison warning that is being treated as an error.

The warning is triggered by a `SizeIs(Ge(int))` matcher in the
`SubscriptionSessionTest.ScheduleCallbacksWithOtelEnabled` test.
The `SizeIs` matcher compares the size of a container (a `size_t`)
with an `int`, which causes a signed/unsigned mismatch warning.

This change fixes the build by casting the integer to `std::size_t`
to avoid the warning.