fix(bazel): bind libcrypto and libssl to boringssl
The macOS Bazel build is failing with an error indicating that the
`grpc_jwt_credentials` target does not have a dependency on a module
exporting OpenSSL headers.

This is because the `-no-canonical-prefixes` flag, which is needed to
avoid OpenSSL header conflicts on macOS, is preventing gRPCs build
system from correctly resolving its `libcrypto` and `libssl` external
dependencies.

This change fixes the build by using `native.bind` to create global
aliases for `libcrypto` and `libssl`, pointing them to the `//:crypto`
and `//:ssl` targets in the `boringssl` repository. This allows gRPCs
build system to correctly resolve these dependencies.